Node.bind(name, observable)
This patch changes the Node.bind so that bind always takes a binding name and an observable. The implication of this is that creating observers is now always the concern of TemplateBinding (and in some case the registered delegate).

Note that NodeBinding is now gone. The observable is now serving the same purpose the NodeBinding was.

R=arv
BUG=

Review URL: https://codereview.appspot.com/40490043