Allow negative conditional binding
There are some use cases where you need a negative conditional binding. If you have a model property visible and want to add an attribute hidden depending on this property, you don't want to add an additional property to the model just for the binding. The same applies for the disabled property of an select box where you have a model property enabled.