Revert support for binding to on-* as event handlers
The current implementation lacks an important feature on which polymer is relying: the receiver of the function is the model (or the object most proximate to the path-value).

It's not clear to me at the moment how to achieve this with the current layering, so I'm reverting until we have a complete design -- as this implementation is interfering with polymers.

R=arv
BUG=

Review URL: https://codereview.appspot.com/48180043